Main/src-server: clang build fixes, bugref:10391
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@108628 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-server/linux/HostDnsServiceLinux.cpp
@@ -161,7 +161,7 @@ static int monitorSymlinkedDir(int iInotifyFd, char szRealResolvConf[PATH_MAX], 
             && strchr(szRealResolvConf, '/'))
         {
             /* Cut of the filename part. We only need that for deletion checks and such. */
-            size_t const offFilename = strrchr(szRealResolvConf, '/') - &szRealResolvConf[0];
+            size_t const offFilename = (size_t)(strrchr(szRealResolvConf, '/') - &szRealResolvConf[0]);
             *poffFilename = offFilename + 1;
             szRealResolvConf[offFilename] = '\0';
